organizing the endpoints

+++ FinAspire.API/Handlers/CategoryHandler.cs
@@ -1,9 +1,9 @@
-﻿using FinAspire.Core.Handler;
+﻿using System.Text;
+using FinAspire.Core.Handler;
 using FinAspire.Core.Models;
 using FinAspire.Core.Request.Categories;
 using FinAspire.Core.Response;
 using FinAspire.Infra.Repositories.Categories;
-using Microsoft.EntityFrameworkCore;
 
 namespace FinAspire.API.Handlers;
 
@@ -17,6 +17,7 @@ public class CategoryHandler(ICategoryRepository repository): ICategoryHandler
             {
                 Title = request.Title,
                 Description = request.Description,
+                UserId = request.UserId
             };
 
             await repository.CreateAsync(category);
@@ -25,7 +26,10 @@ public class CategoryHandler(ICategoryRepository repository): ICategoryHandler
         }
         catch (Exception e)
         {
-            return new BaseResponse<Category>(code: 500, message: e.Message, data: null);
+            var sb = new StringBuilder();
+            sb.Append("[CAT003] ");
+            sb.Append(e.Message);
+            return new BaseResponse<Category>(code: 500, message: sb.ToString(), data: null);
         }
     }
 
@@ -33,19 +37,26 @@ public class CategoryHandler(ICategoryRepository repository): ICategoryHandler
     {
         try
         {
-            var category = new Category
+            var category = await repository.GetByIdAsync(request.Id, request.UserId);
+
+            if (category == null)
             {
-                Id = request.Id,
-                Title = request.Title ?? String.Empty,
-                Description = request.Description,
-            };
+                Console.WriteLine("[CAT001] Category not found");
+                return new BaseResponse<Category>(code: 204, message: "[CAT001] Category not found", data: null);
+            }
+            
+            category.Title = request.Title;
+            category.Description = request.Description;
+            
             var result = await repository.UpdateAsync(category);
             return new BaseResponse<Category>(result, message: "Category updated successfully", code: 200);
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            throw;
+            var sb = new StringBuilder();
+            sb.Append("[CAT002] ");
+            sb.Append(e.Message);
+            return new BaseResponse<Category>(code: 500, message: sb.ToString(), data: null);
         }
     }
 
@@ -53,6 +64,11 @@ public class CategoryHandler(ICategoryRepository repository): ICategoryHandler
     {
         try
         {
+            var category = await repository.GetByIdAsync(request.Id, request.UserId);
+            
+            if (category == null) 
+                return new BaseResponse<Category>(code: 204, message: "[CAT003] Category not found", data: null);
+            
             var result = await repository.DeleteAsync(request.Id);
             
             return new BaseResponse<Category>(result, message: "Category deleted successfully", code: 200);
@@ -60,8 +76,10 @@ public class CategoryHandler(ICategoryRepository repository): ICategoryHandler
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            return new BaseResponse<Category>(code: 500, message: e.Message, data: null);
+            var sb = new StringBuilder();
+            sb.Append("[CAT004] ");
+            sb.Append(e.Message);
+            return new BaseResponse<Category>(code: 500, message: sb.ToString(), data: null);
         }
     }
 
@@ -69,30 +87,36 @@ public class CategoryHandler(ICategoryRepository repository): ICategoryHandler
     {
         try
         {
-            var result = await repository.GetByIdAsync(request.Id);
+            var result = await repository.GetByIdAsync(request.Id, request.UserId);
             
             return result == null ? new BaseResponse<Category>(code: 204, message: "Category not found", data: null) 
                 : new BaseResponse<Category>(result, message: "Category found", code: 200);
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            throw;
+            var sb = new StringBuilder();
+            sb.Append("[CAT005] ");
+            sb.Append(e.Message);
+            return new BaseResponse<Category>(code: 500, message: sb.ToString(), data: null);
         }
     }
 
-    public async Task<BaseResponse<List<Category>>> GetAllAsync(GetAllCategoriesRequest request)
+    //TODO Rever esse metodo, não esta fazendo muito sentido trafegar o request
+    public async Task<PagedResponse<List<Category>?>> GetAllAsync(GetAllCategoriesRequest request)
     {
         try
         {
-            var result = await repository.GetAllAsync();
+            var result = await repository.GetAllAsync(request);
+            result.Message = "All categories found";
             
-            return new BaseResponse<List<Category>>(result.ToList(), message: "Categories found", code: 200);
+            return result;
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            throw;
+            var sb = new StringBuilder();
+            sb.Append("[CAT006] ");
+            sb.Append(e.Message);
+            return new PagedResponse<List<Category>?>(code: 500, message: sb.ToString(), data: null);
         }
     }
 }
